requirements.txt
fastapi
pydantic
anthropic
python-dotenv
uvicorn
